django_API_intermidiate/4-Django-Rest-Framework/requirements.txt
argon2-cffi==19.1.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cx-Oracle==7.2.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.3
django-admin-sortable==2.1.17
django-allauth==0.39.1
django-bootstrap3==11.1.0
django-braces==1.13.0
django-cors-headers==3.0.2
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
Markdown==3.1.1
misaka==2.1.1
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==6.1.0
protobuf==3.8.0
psycopg2==2.8.3
pycparser==2.19
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.4
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
wincertstore==0.2
